Move the Split Streams execution to the database, so it can more easily re-use results. (Might need to verify whether this actually improves things?)
